requirements.txt
Django>=6.0
python-dotenv
boto3
agno-ai
gunicorn
